+++ pytodotxt/todotxt.py
@@ -1,56 +1,72 @@
-import datetime
-import re
+"""TodoTxt class and parser
+"""
 import pathlib
-import os
 import tempfile
+import os
+import io
+
+from .task import Task
 
 
 class TodoTxt:
-    """Access to a todo.txt file
+    """Convenience wrapper for a single todo.txt file
 
-    Common use::
+    The most common use is::
 
         todotxt = TodoTxt("todo.txt")
         todotxt.parse()
 
     Use the ``tasks`` property to access the parsed entries.
     """
-    def __init__(self, filename, encoding='utf-8', task_class=None):
+    def __init__(self, filename, encoding='utf-8', parser=None):
+        """Create a new TodoTxt instance from ``filename``
+
+        :param filename: the file you wish to use
+        :type filename: ``str`` or ``pathlib.Path``
+        :param encoding: what encoding to assume for the content of the file
+        :param parser: This may be a custom parser instance to use
+                       instead of the default ``TodoTxtParser``.
+        :type parser: ``TodoTxtParser``
+        """
         self.filename = pathlib.Path(filename)
         self.encoding = encoding
         self.linesep = os.linesep
-        self.task_class = Task if task_class is None else task_class
+
         self.tasks = []
+        """The tasks of this file"""
+
+        self.parser = parser or TodoTxtParser(self.encoding)
+
+    def add(self, task):
+        """Add a task to this container
+
+        You could also just ``append`` to ``self.tasks``,
+        but this function call will also update the ``todotxt`` and
+        ``linenr`` properties of ``task``.
+
+        :param task: The task that should be added to this todo.txt file
+        :type task: ``Task``
+        """
+        self.tasks.append(task)
+        task.linenr = len(self.tasks)
+        task.todotxt = self
 
     def parse(self):
-        """(Re)parse the todo.txt file"""
+        """(Re)parse ``self.filename``
+
+        This will try to detect the line separator of the file automatically
+        and remember it as ``self.linesep`` for the time when you save the
+        file back to disk.
+
+        :return: the list of all tasks read from the file.
+        """
         self.tasks = []
 
-        # process task lines of file
-        with open(self.filename, 'rt', encoding=self.encoding) as fd:
-            lines = fd.readlines()
+        for task in self.parser.parse(pathlib.Path(self.filename)):
+            task.todotxt = self
+            self.tasks.append(task)
 
-            # remember newline separator
-            if isinstance(fd.newlines, str):
-                self.linesep = fd.newlines
-            # handle the case when multiple newline separators are detected
-            elif isinstance(fd.newlines, tuple):
-                # when the OS newline seperator has been used in the source file
-                if os.linesep in fd.newlines:
-                    # use the system default newline separator
-                    self.linesep = os.linesep
-                # otherwise if the newline seperator of the OS has not been used in the source file but other ones
-                else:
-                    # use the first found newline separator
-                    self.linesep = fd.newlines[0]
-
-            # read lines and parse them as tasks
-            for linenr, line in enumerate(lines):
-                if len(line.strip()) == 0:
-                    continue
-                task = self.task_class(line, linenr=linenr, todotxt=self)
-                self.tasks.append(task)
-
+        self.linesep = self.parser.linesep
         return self.tasks
 
     def save(self, target=None, safe=True, linesep=None):
@@ -66,322 +82,141 @@ class TodoTxt:
         some cloud storage.
 
         With ``linesep`` you can specify the line seperator. If it is not set
-        it defaults to the systems default line seperator.
+        it defaults to the systems default line seperator (or the detected
+        line separator, if the file has been parsed from disk first).
+
+        :param target: The file to save to, or ``None`` to use ``self.filename``
+        :param safe: Whether or not to save the file in a way that does not
+                     destroy the previous file in case on errors.
+        :param linesep: The line separator, or ``None`` to use ``self.linesep``
         """
         if target is None:
             target = self.filename
         else:
             target = pathlib.Path(target)
-        write_to = target
-
-        tmpfile = None
-        if safe:
-            tmpfile = tempfile.NamedTemporaryFile(dir=self.filename.parent,
-                                                  delete=False,
-                                                  prefix=".tmp",
-                                                  suffix="~")
-            write_to = tmpfile.name
-            tmpfile.close()
 
         if linesep is None:
             linesep = self.linesep
 
-        with open(write_to, 'wb', buffering=0) as fd:
-            lines = [str(task) + linesep for task in
-                     sorted(self.tasks, key=lambda t: t.linenr if t.linenr is not None else len(self.tasks))]
-            fd.write(bytes(''.join(lines), self.encoding))
-
         if safe:
-            os.replace(write_to, target)
+            tmpfile = tempfile.NamedTemporaryFile('wb',
+                                                  buffering=0,
+                                                  dir=self.filename.parent,
+                                                  delete=False,
+                                                  prefix=".tmp",
+                                                  suffix="~")
+            self.write_to_stream(tmpfile, linesep)
+            tmpfile.close()
 
+            os.replace(tmpfile.name, target)
             try:
-                os.unlink(write_to)
+                os.unlink(tmpfile.name)
             except OSError:
                 pass
+        else:
+            with open(target, 'wb', buffering=0) as fd:
+                self.write_to_stream(fd, linesep)
+
+    def write_to_stream(self, stream, linesep=None):
+        """Save all tasks, ordered by their line number, to the stream
+
+        :param stream: an ``io.IOBase`` like stream that accepts bytes being
+                       written to it.
+        :param linesep: the line separator, or ``None`` to use ``self.linesep``
+        """
+        if linesep is None:
+            linesep = self.linesep
+        stream.write(bytes(linesep.join(self.lines), self.encoding))
+
+    @property
+    def lines(self):
+        """Much like ``self.tasks``, but already sorted by linenr and converted to string"""
+        lines = [(task.linenr if task.linenr is not None else len(self.tasks), str(task))
+                 for task in self.tasks]
+        lines.sort()
+        return [line for _, line in lines]
 
     def __repr__(self):
         return f'{self.__class__.__name__}(filename="{self.filename}")'
 
 
-class Task:
-    """A task of a todo.txt file
+class TodoTxtParser:
+    """A parser for todo.txt-like formatted strings or files"""
+    def __init__(self, encoding='utf-8', task_type=None):
+        """Create a new todo.txt parser
 
-    The usual way to create a task is to create it from an initial string::
+        :param encoding: The encoding to assume for the parsing process
+        :param task_type: A subclass of ``Task`` to use when parsing the tasks."""
+        self.encoding = encoding
+        self.task_type = task_type or Task
 
-        task = Task("(B) some task")
+        self.linesep = os.linesep
+        """The line separator. After running any of the parse functions, this
+        property will be set to the line separator that was detected in the
+        parsed object."""
 
-    or::
+    def parse(self, target):
+        """Parse the given object
 
-        task = Task()
-        task.parse("(B) some task")
+        The behaviour of the parser depends a bit on what you pass in.
+        It might be a bit surprising, but if you pass in a ``str``, ``parse``
+        will attempt to find tasks in that string.
 
-    The inverse operation to parsing is to convert the task to a string::
+        If you want to parse a file and pass the filename, wrap it in
+        ``pathlib.Path`` first or use ``parse_file`` directly.
 
-        task = Task("(B) some task")
-        assert str(task) == "(B) some task"
+        When parsing is completed, you can query ``linesep`` to see what
+        the line separator was.
 
-    """
-    COMPLETED_RE = re.compile(r'^x\s+')
-    PRIORITY_RE = re.compile(r'^\s*\(([A-Z]+)\)')
-    PROJECT_RE = re.compile(r'(\s+|^)\+([^\s]+)')
-    CONTEXT_RE = re.compile(r'(\s+|^)@([^\s]+)')
-    KEYVALUE_RE = re.compile(r'(\s+|^)([^\s]+):([^\s$]+)')
-    DATE_RE = re.compile(r'^\s*([\d]{4}-[\d]{2}-[\d]{2})', re.ASCII)
-    DATE_FMT = '%Y-%m-%d'
-    KEYVALUE_ALLOW = set(['http', 'https', 'mailto', 'ssh', 'ftp'])
-
-    def __init__(self, line=None, linenr=None, todotxt=None):
-        """Create a new task
-
-        ``line`` is the raw string representation (one line of a todo.txt file).
-        ``linenr`` is the line number within the ``todotxt`` file, if any.
+        :param target: the object to parse
+        :type target: ``pathlib.Path``, ``str``, ``bytes``, or any ``io.IOBase``.
+        :return: a list of tasks found in ``target``
         """
-        self.description = None
-        self.is_completed = None
-        self.priority = None
-        self.completion_date = None
-        self.creation_date = None
-        self.linenr = linenr
-        self.todotxt = todotxt
-        self._raw = None
-        self._attributes = None
+        if isinstance(target, pathlib.Path):
+            return self.parse_file(target)
+        if isinstance(target, str):
+            return self.parse_str(target)
+        if isinstance(target, bytes):
+            return self.parse_str(str(target, self.encoding))
+        if isinstance(target, io.IOBase):
+            return self.parse_stream(target)
+        raise TypeError(f"Don't know how to parse {type(target).__name__}")
 
-        if line is not None and len(line.strip()) > 0:
-            self.parse(line)
+    def parse_str(self, text):
+        """Parse tasks from the given text
 
-    def remove_project(self, project):
-        return self.remove_tag(project, self.__class__.PROJECT_RE)
-
-    def remove_context(self, context):
-        return self.remove_tag(context, self.__class__.CONTEXT_RE)
-
-    def remove_attribute(self, key, value=None):
-        """Remove attribute `key` from the task
-        If you provide a `value` only the attribute with that key and value is removed.
-        If no `value` is provided all attributes with that key are removed.
+        :param text: The string containing todo.txt-like tasks
+        :return: a list of tasks found in ``text``
         """
-        success = False
-        while True:
-            key_found = False
+        self.linesep = os.linesep
 
-            for match in self.parse_tags(self.__class__.KEYVALUE_RE):
-                if key != match.group(2):
-                    continue
-
-                key_found = True
-
-                if value is None or match.group(3) == value:
-                    start, end = match.span()
-                    self.description = self.description[:start]
-                    self.parse(str(self))
-                    success = True
-
+        # determine line separator
+        for ch in ['\r\n', '\n', '\r']:
+            if ch in text:
+                self.linesep = ch
                 break
 
-            if not key_found:
-                break
+        return [self.task_type(line, linenr)
+                for linenr, line in enumerate(text.rstrip().split(self.linesep))]
 
-        return success
+    def parse_file(self, path):
+        """Parse the content of the file at ``path``
 
-    def remove_tag(self, text, regex):
-        for match in self.parse_tags(regex):
-            if match.group(2) == text:
-                start, end = match.span()
-                self.description = self.description[:start] + self.description[end:]
-                self.parse(str(self))
-                return True
-        return False
-
-    def replace_attribute(self, key, value, newvalue):
-        """Replace the value of key:value in place with key:newvalue"""
-        for match in self.parse_tags(self.__class__.KEYVALUE_RE):
-            if key != match.group(2) or value != match.group(3):
-                continue
-
-            self.description = self.description[:match.start(3)] + \
-                               newvalue + \
-                               self.description[match.end(3):]
-            self.parse(str(self))
-            return True
-
-        return False
-
-    def replace_context(self, context, newcontext):
-        """Replace the first occurrence of @context with @newcontext"""
-        return self.replace_tag(context, newcontext, self.__class__.CONTEXT_RE)
-
-    def replace_project(self, project, newproject):
-        """Replace the first occurrence of @project with @newproject"""
-        return self.replace_tag(project, newproject, self.__class__.PROJECT_RE)
-
-    def replace_tag(self, value, newvalue, regex):
-        for match in self.parse_tags(regex):
-            if match.group(2) == value:
-                self.description = self.description[:match.start(2)] + \
-                                   newvalue + \
-                                   self.description[match.end(2):]
-                self.parse(str(self))
-                return True
-        return False
-
-    def add_project(self, project):
-        self.append('+' + project)
-        self.parse(str(self))
-
-    def add_context(self, context):
-        self.append('@' + context)
-        self.parse(str(self))
-
-    def add_attribute(self, key, value):
-        self.append(f'{key}:{value}')
-        self.parse(str(self))
-
-    def append(self, text, add_space=True):
-        if self.description is None:
-            self.description = text
-        else:
-            if add_space and not self.description.endswith(' '):
-                self.description += ' '
-            self.description += text
-
-    def bare_description(self):
-        """The description of the task without contexts, projects or any other attributes"""
-        if self.description is None:
-            return ''
-
-        parts = []
-        for part in self.description.split(' '):
-            if len(part) == 0 or part[0] in '@+':
-                continue
-            # make sure attributes with keys in KEYVALUE_ALLOW are included in bare description
-            elif ':' in part:
-                attrribute_key = part[:part.index(":")]
-                if attrribute_key.lower() in self.__class__.KEYVALUE_ALLOW:
-                    parts.append(part)
-            else:
-                parts.append(part)
-
-        return ' '.join(parts)
-
-    @property
-    def projects(self):
-        return [match.group(2) for match in self.parse_tags(self.__class__.PROJECT_RE)]
-
-    @property
-    def contexts(self):
-        return [match.group(2) for match in self.parse_tags(self.__class__.CONTEXT_RE)]
-
-    @property
-    def attributes(self):
-        if self._attributes is None:
-            self.parse_attributes()
-        return self._attributes
-
-    def __getattr__(self, name, fallback=None):
-        if name.startswith('attr_'):
-            if fallback is None:
-                fallback = []
-            _, attrname = name.split('_', 1)
-            return self.attributes.get(attrname, fallback)
-        raise AttributeError(name)
-
-    def parse_attributes(self):
-        self._attributes = {}
-        for match in self.parse_tags(self.__class__.KEYVALUE_RE):
-            key = match.group(2)
-            value = match.group(3)
-            if key.lower() in self.__class__.KEYVALUE_ALLOW:
-                continue
-            if key not in self._attributes:
-                self._attributes[key] = []
-            self._attributes[key].append(value)
-
-    def parse_tags(self, regex):
-        matches = []
-        if self.description is None:
-            return matches
-
-        for match in regex.finditer(self.description):
-            if match:
-                matches.append(match)
-            else:
-                break
-        return matches
-
-    def parse_priority(self, line):
-        self.priority = None
-        match = self.__class__.PRIORITY_RE.match(line)
-        if match:
-            self.priority = match.group(1)
-            line = line[match.span()[1]:]
-        return line
-
-    @classmethod
-    def match_date(cls, line):
-        match = cls.DATE_RE.match(line)
-        date = None
-        if match:
-            date = cls.parse_date(match.group(1))
-            line = line[match.span()[1]:]
-        return line, date
-
-    @classmethod
-    def parse_date(cls, text):
-        return datetime.datetime.strptime(text, cls.DATE_FMT).date()
-
-    def parse(self, line):
-        """(Re)parse the task
-
-        ``line`` is the raw string representation of a task, i.e. one line
-        of a todo.txt file.
+        :param path: The file to parse
+        :return: a list of tasks found in ``file``.
         """
-        self._raw = line
-        line = line.strip()
+        with open(path, 'rt', encoding=self.encoding) as fh:
+            return self.parse_stream(fh)
 
-        # completed or not
-        match = self.__class__.COMPLETED_RE.match(line)
-        self.is_completed = match is not None
-        if match:
-            # strip the leading mark
-            line = line[match.span()[1]:]
+    def parse_stream(self, stream):
+        """Parse an io stream
 
-        if self.is_completed:
-            line, self.completion_date = self.__class__.match_date(line)
+        :param stream: an io stream to read and parse tasks from
+        :return: a list of tasks found in ``stream``
+        """
+        # get the text from the stream
+        blob = stream.read()
+        if not isinstance(blob, str):
+            blob = str(blob, self.encoding)
 
-        line = self.parse_priority(line)
-        line, self.creation_date = self.__class__.match_date(line)
-
-        # description
-        if len(line) > 0:
-            self.description = line.strip()
-
-        self.parse_attributes()
-
-    def __str__(self):
-        """The todo.txt compatible representation of this task."""
-        result = ''
-        if self.is_completed:
-            result += 'x '
-
-            if self.completion_date is not None:
-                result += self.completion_date.strftime(self.__class__.DATE_FMT) + ' '
-
-        if self.priority:
-            if not self.is_completed:
-                result += f'({self.priority}) '
-
-        if self.creation_date is not None:
-            result += self.creation_date.strftime(self.__class__.DATE_FMT) + ' '
-
-        if self.description:
-            result += self.description
-
-        return result
-
-    def __repr__(self):
-        return f'{self.__class__.__name__}({repr(str(self))})'
-
-
-# for backwards compatability
-match_date = Task.match_date
-parse_date = Task.parse_date
+        return self.parse_str(blob)